Fix/#930 shrinking input (#933)
* Replaced not existing style attribute for placeholder

There was a style attribute referring to a not existing style, this has been replaced for a placeholder attribute.

* Fixed bug at Subscriptions where long string would make the input shrink

When given a string that was too long, the input checkbox on the end would shrink. This has been fixed by giving it a min-width. On the mobile view the last char would get chopped of, for this reason also the char limit has been decreased by one.

closes #930